LPD-48736 Merge them into one class
