+++ modules/apps/portal-osgi-debug/portal-osgi-debug-impl/src/main/resources/META-INF/module-log4j.xml
@@ -2,6 +2,6 @@
 
 <Configuration strict="true">
 	<Loggers>
-		<Logger level="INFO" name="com.liferay.portal.osgi.debug.internal.osgi.commands.SystemCheckOSGiCommands" />
+		<Logger level="INFO" name="com.liferay.portal.osgi.debug.internal.osgi.commands.SystemOSGiCommands" />
 	</Loggers>
 </Configuration>
